Move ExpressionList to list.generic
