+++ clock/storage/data_source/data_sources/sqlite/sql/item/expression/compound/condition.py
@@ -1,5 +1,5 @@
 from clock.storage.data_source.data_sources.sqlite.sql.item.constants.operator import Operator
-from clock.storage.data_source.data_sources.sqlite.sql.item.expression.compound.list import ExpressionList
+from clock.storage.data_source.data_sources.sqlite.sql.item.expression.compound.list.generic import ExpressionList
 from clock.storage.data_source.data_sources.sqlite.sql.item.expression.parser import EXPRESSION_TYPE
